@@ -13,6 +13,7 @@ import { AuthModule} from './auth/auth.module';
 import { ManageModule } from './manage/manage.module';
 import { UpperCasePipe } from './common/pipes/uppercase.pipe';
 
+
 const routes: Routes = [
   {path: '', redirectTo: '/rentals', pathMatch: 'full'},
 ]
@@ -22,7 +23,6 @@ const routes: Routes = [
     AppComponent,
     HeaderComponent,
     UpperCasePipe
-   
   ],
   imports: [
     RouterModule.forRoot(routes),
@@ -31,10 +31,9 @@ const routes: Routes = [
     RentalModule,
     AuthModule,
     ManageModule,
-    NgbModule.forRoot(),    
+    NgbModule.forRoot(),
     BrowserAnimationsModule,
-    ToasterModule.forRoot(),
-   
+    ToasterModule.forRoot()
   ],
   providers: [],
   bootstrap: [AppComponent]
