@@ -5,15 +5,12 @@ import { AuthService } from './auth.service';
 @Injectable()
 export class AuthGuardService implements CanActivate {
     private url: string;
-
-  
     constructor(
-        private authService: AuthService, 
+        private authService: AuthService,
         public router: Router
     ) {}
-    
     private handleAuthState(): boolean {
-        if(this.isLoginOrRegisterPage()){
+        if (this.isLoginOrRegisterPage()) {
             this.router.navigate(['/rentals']);
             return false;
         }
@@ -21,7 +18,7 @@ export class AuthGuardService implements CanActivate {
     }
 
     private handleNotAuthState(): boolean {
-        if(this.isLoginOrRegisterPage()){
+        if (this.isLoginOrRegisterPage()) {
             return true;
         }
         this.router.navigate(['/login']);
@@ -29,14 +26,15 @@ export class AuthGuardService implements CanActivate {
     }
 
     private isLoginOrRegisterPage(): boolean {
-        if(this.url.includes('login') || this.url.includes('register'))
-            return true;
+        if (this.url.includes('login') || this.url.includes('register')) {
+          return true;
+        }
         return false;
     }
 
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot ): boolean {
         this.url = state.url;
-        if(this.authService.isAuthentificated()){
+        if (this.authService.isAuthentificated()) {
             return this.handleAuthState();
         }
 
