@@ -1,5 +1,5 @@
-import { Component, OnInit, Input, ChangeDetectorRef } from '@angular/core';
-
+import {Component, OnInit, Input, ChangeDetectorRef, OnDestroy} from '@angular/core';
+import { Subject } from 'rxjs';
 import { MapService} from './map.service';
 
 @Component({
@@ -7,9 +7,11 @@ import { MapService} from './map.service';
   templateUrl: './map.component.html',
   styleUrls: ['./map.component.scss']
 })
-export class MapComponent implements OnInit {
+export class MapComponent implements OnInit, OnDestroy {
 
   @Input() location: string;
+  @Input() locationSubject: Subject<any>;
+
   lat: number = 0;
   lng: number = 0;
 
@@ -21,10 +23,25 @@ export class MapComponent implements OnInit {
     ) {}
 
   ngOnInit() {
+    if (this.locationSubject) {
+      this.locationSubject.subscribe((location: string) => {
+        this.getLocation(location);
+      });
+    }
   }
 
-  mapReadyHandler(){
-    this.mapService.getGeoLocation(this.location).subscribe(
+  ngOnDestroy(): void {
+    if (this.locationSubject) {
+      this.locationSubject.unsubscribe();
+    }
+  }
+
+  mapReadyHandler() {
+    this.getLocation(this.location);
+  }
+
+  private getLocation(location) {
+    this.mapService.getGeoLocation(location).subscribe(
       (coordinates) => {
         this.lat = coordinates.lat;
         this.lng = coordinates.lng;
@@ -33,11 +50,13 @@ export class MapComponent implements OnInit {
         console.log(this.lat, this.lng);
 
         this.changeDetectorRef.detectChanges();
-      }, () =>{
+      }, () => {
         this.isPositionError = true;
         console.log('Error getting coords from GeoServer');
         this.changeDetectorRef.detectChanges();
       }
     );
   }
+
+
 }
